Separate plugins/init.lua file into multiple files
The separation into multiple files was only done for plugins that
required more than a single line to import.
