+++ lua/plugins/refactoring.lua
@@ -1,0 +1,11 @@
+return {  -- Adds several more refactoring options.
+  'ThePrimeagen/refactoring.nvim',
+  dependencies = {
+    'nvim-lua/plenary.nvim',
+    'nvim-treesitter/nvim-treesitter',
+  },
+  lazy = false,
+  config = function()
+    require('refactoring').setup {}
+  end,
+}
